+++ queryparams/params.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"net/url"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/ONSdigital/log.go/v2/log"
@@ -252,219 +251,6 @@ func CalculatePageNumber(offset, pageSize int) int {
 	}
 
 	return ((offset + 1) / pageSize) + 1
-}
-
-type Sort int
-
-const (
-	invalidSort Sort = iota
-	RelDateAsc
-	RelDateDesc
-	TitleAZ
-	TitleZA
-	Relevance
-)
-
-var sortValues = map[Sort]struct{ feValue, beValue string }{
-	RelDateAsc:  {feValue: "date-oldest", beValue: "release_date_asc"},
-	RelDateDesc: {feValue: "date-newest", beValue: "release_date_desc"},
-	TitleAZ:     {feValue: "alphabetical-az", beValue: "title_asc"},
-	TitleZA:     {feValue: "alphabetical-za", beValue: "title_desc"},
-	Relevance:   {feValue: "relevance", beValue: "relevance"},
-}
-
-func parseSort(sort string) (Sort, error) {
-	for s, sv := range sortValues {
-		if strings.EqualFold(sort, sv.feValue) {
-			return s, nil
-		}
-	}
-
-	return invalidSort, errors.New("invalid sort option string")
-}
-
-func (s Sort) String() string {
-	return sortValues[s].feValue
-}
-
-func (s Sort) BackendString() string {
-	return sortValues[s].beValue
-}
-
-type Date struct {
-	date       time.Time
-	y, m, d    int
-	ys, ms, ds string
-}
-
-const DateFormat = "2006-01-02"
-
-func MustParseDate(dateAsString string) Date {
-	d, err := ParseDate(dateAsString)
-	if err != nil {
-		panic("invalid date string: " + dateAsString)
-	}
-
-	return d
-}
-
-func ParseDate(dateAsString string) (Date, error) {
-	if dateAsString == "" {
-		return Date{}, nil
-	}
-	t, err := time.Parse(DateFormat, dateAsString)
-	if err != nil {
-		return Date{}, err
-	}
-
-	return DateFromTime(t), nil
-}
-
-func DateFromTime(t time.Time) Date {
-	if t.IsZero() {
-		return Date{}
-	}
-	date := Date{date: t}
-	y, m, d := t.Date()
-	date.y, date.m, date.d = y, int(m), d
-	date.ys, date.ms, date.ds = strconv.Itoa(y), strconv.Itoa(int(m)), strconv.Itoa(d)
-
-	return date
-}
-
-func (d Date) String() string {
-	if d.date.IsZero() {
-		return ""
-	}
-
-	return d.date.UTC().Format(DateFormat)
-}
-
-func (d Date) YearString() string {
-	return d.ys
-}
-
-func (d Date) MonthString() string {
-	return d.ms
-}
-
-func (d Date) DayString() string {
-	return d.ds
-}
-
-type ReleaseType int
-
-const (
-	invalidReleaseType ReleaseType = iota
-	Upcoming
-	Published
-	Cancelled
-	Provisional
-	Confirmed
-	Postponed
-)
-
-var relTypeValues = map[ReleaseType]struct{ name, label string }{
-	Upcoming:    {name: "type-upcoming", label: "Upcoming"},
-	Published:   {name: "type-published", label: "Published"},
-	Cancelled:   {name: "type-cancelled", label: "Cancelled"},
-	Provisional: {name: "subtype-provisional", label: "Provisional"},
-	Confirmed:   {name: "subtype-confirmed", label: "Confirmed"},
-	Postponed:   {name: "subtype-postponed", label: "Postponed"},
-}
-
-func parseReleaseType(s string) (ReleaseType, error) {
-	for rt, rtv := range relTypeValues {
-		if strings.EqualFold(s, rtv.name) {
-			return rt, nil
-		}
-	}
-
-	return invalidReleaseType, errors.New("invalid release type string")
-}
-
-func (rt ReleaseType) Name() string {
-	return relTypeValues[rt].name
-}
-
-func (rt ReleaseType) Label() string {
-	return relTypeValues[rt].label
-}
-
-func (rt ReleaseType) String() string {
-	return rt.Name()
-}
-
-type ValidatedParams struct {
-	Limit       int
-	Page        int
-	Offset      int
-	AfterDate   Date
-	BeforeDate  Date
-	Keywords    string
-	Sort        Sort
-	ReleaseType ReleaseType
-	Provisional bool
-	Confirmed   bool
-	Postponed   bool
-	Census      bool
-	Highlight   bool
-}
-
-// AsBackendQuery converts to a url.Values object with parameters as expected by the api
-func (vp ValidatedParams) AsBackendQuery() url.Values {
-	return vp.asQuery(true)
-}
-
-// AsFrontendQuery converts to a url.Values object with parameters
-func (vp ValidatedParams) AsFrontendQuery() url.Values {
-	return vp.asQuery(false)
-}
-
-func (vp ValidatedParams) asQuery(isBackend bool) url.Values {
-	var query = make(url.Values)
-	setValue(query, Limit, strconv.Itoa(vp.Limit))
-	setValue(query, Page, strconv.Itoa(vp.Page))
-
-	if isBackend {
-		setValue(query, Offset, strconv.Itoa(vp.Offset))
-		setValue(query, Query, vp.Keywords)
-		setValue(query, SortName, vp.getSortBackendString())
-		setValue(query, DateFrom, vp.AfterDate.String())
-		setValue(query, DateTo, vp.BeforeDate.String())
-	} else {
-		setValue(query, Keywords, vp.Keywords)
-		setValue(query, SortName, vp.Sort.String())
-		setValue(query, YearBefore, vp.BeforeDate.YearString())
-		setValue(query, MonthBefore, vp.BeforeDate.MonthString())
-		setValue(query, DayBefore, vp.BeforeDate.DayString())
-		setValue(query, YearAfter, vp.AfterDate.YearString())
-		setValue(query, MonthAfter, vp.AfterDate.MonthString())
-		setValue(query, DayAfter, vp.AfterDate.DayString())
-	}
-
-	setValue(query, Type, vp.ReleaseType.String())
-	if vp.ReleaseType == Upcoming {
-		setBoolValue(query, Provisional.String(), vp.Provisional)
-		setBoolValue(query, Confirmed.String(), vp.Confirmed)
-		setBoolValue(query, Postponed.String(), vp.Postponed)
-	}
-	setBoolValue(query, Census, vp.Census)
-	setBoolValue(query, Highlight, vp.Highlight)
-
-	return query
-}
-
-func (vp ValidatedParams) getSortBackendString() string {
-	// Newest is now defined as 'the closest date to today' and so its
-	// meaning in terms of algorithmic definition (ascending/descending)
-	// is reversed depending on the release-type that is being viewed
-	if vp.ReleaseType == Upcoming && vp.Sort == RelDateDesc {
-		return RelDateAsc.BackendString()
-	} else if vp.ReleaseType == Upcoming && vp.Sort == RelDateAsc {
-		return RelDateDesc.BackendString()
-	}
-	return vp.Sort.BackendString()
 }
 
 func setValue(query url.Values, key string, value string) {
